# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ICOS Job Manager Microservice.",
        "title": "Swagger Job Manager API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.4.1-latest"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
...
        "/jobmanager/resources/status": {
            "put": {
                "description": "update resource status by uuid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Update resource status by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resource UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Resource info",
                        "name": "resource",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Resource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Resource UUID is required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Can not find Resource to update",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/jobmanager/resources/status/{job_uuid}": {
            "get": {
                "description": "get resource status by job uuid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Get resource status by job UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job UUID",
                        "name": "job_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Resource"
                        }
                    },
                    "400": {
                        "description": "Job UUID is required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Can not find Job by UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Condition": {
            "type": "object",
            "required": [
                "lastTransitionTime",
                "message",
                "reason",
                "status",
                "type"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastTransitionTime": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "observedGeneration": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.ConditionStatus"
                },
                "type": {
                    "$ref": "#/definitions/models.ResourceState"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ConditionStatus": {
            "type": "string",
            "enum": [
                "True",
                "False",
                "Unknown"
            ],
            "x-enum-varnames": [
                "ConditionTrue",
                "ConditionFalse",
                "ConditionUnknown"
            ]
        },
...
        "models.Resource": {
            "type": "object",
            "properties": {
                "conditions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Condition"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "job_id": {
                    "type": "string"
                },
                "resource_name": {
                    "type": "string"
                },
                "resource_uuid": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ResourceState": {
            "type": "string",
            "enum": [
                "Progressing",
                "Applied",
                "Available",
                "Degraded"
            ],
            "x-enum-varnames": [
                "Progressing",
                "Applied",
                "Available",
                "Degraded"
            ]
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}